requirements.txt
# Core vision
opencv-python>=4.8,<5
mediapipe>=0.10.8,<0.11
numpy>=1.24,<3

# UI
PySide6>=6.5

# Cursor control
pyautogui>=0.9

# AI word correction
pyspellchecker>=0.7
requests>=2.28

# Cloud / Firebase
firebase-admin>=6.0